Inline the CheckDatabase task at use sites
Also extracts the usage in DeckPicker in a new method to avoid bulking up the calling method
