+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionOperations.kt
@@ -421,3 +421,20 @@ fun changeDeckMulti(
         return@executeInTransaction Computation.ok(cards)
     }
 }
+
+fun checkDatabase(
+    col: Collection,
+    collectionTask: ProgressSenderAndCancelListener<String>,
+): Pair<Boolean, Collection.CheckDatabaseResult?> {
+    Timber.d("doInBackgroundCheckDatabase")
+    // Don't proceed if collection closed
+    val result = col.fixIntegrity(TaskManager.ProgressCallback(collectionTask, AnkiDroidApp.appResources))
+    return if (result.failed) {
+        // we can fail due to a locked database, which requires knowledge of the failure.
+        Pair(false, result)
+    } else {
+        // Close the collection and we restart the app to reload
+        CollectionHelper.instance.closeCollection(true, "Check Database Completed")
+        Pair(true, result)
+    }
+}
