+++ AnkiDroid/src/test/java/com/ichi2/async/CheckDatabaseTest.kt
@@ -17,7 +17,7 @@ package com.ichi2.async
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.ichi2.anki.RobolectricTest
-import com.ichi2.async.CollectionTask.CheckDatabase
+import com.ichi2.libanki.Collection
 import com.ichi2.testutils.CollectionUtils
 import org.hamcrest.MatcherAssert.assertThat
 import org.hamcrest.Matchers.equalTo
@@ -30,7 +30,11 @@ class CheckDatabaseTest : RobolectricTest() {
     @Test
     fun checkDatabaseWithLockedCollectionReturnsLocked() {
         lockDatabase()
-        val result = CheckDatabase().execTask(col, mock())
+        val result = object : TaskDelegate<String, Pair<Boolean, Collection.CheckDatabaseResult?>>() {
+            override fun task(col: Collection, collectionTask: ProgressSenderAndCancelListener<String>): Pair<Boolean, Collection.CheckDatabaseResult?> {
+                return checkDatabase(col, collectionTask)
+            }
+        }.execTask(col, mock())
         assertThat("The result should specify a failure", result.first, equalTo(false))
         val checkDbResult = result.second!!
         assertThat("The result should specify the database was locked", checkDbResult.databaseLocked)
